Added header-bar component
The header bar component now renders the header-bar from the dhis2 core.
It contains most of the same functionality as the angular version that is used
in tracker capture. The code included in this project is however a slightly
modified version of the core one. This one should be refactored and be the
primary version to use.