[FormBuilder] Adjustments for synchronous validators and onBlur fields
Synchronous validators are now executed for every onChange event even for onBlur fields,
which means that only asynchronous validators and the onUpdateField callback behave
differently for onBlur fields.

Also fix a bug that caused the onUpdateField callback to be called for onBlur fields
even when the value didn't actually change.