[FormBuilder] Adjust for possible async setState
Reacts setState is not always synchronous. Because we call
setState fairly often in the handleFieldChange method, these
operations do not always happen synchronous which makes the
state updates happen out of order or be swallowed by Reacts
batching mechanism. Current implementation is not ideal as it
makes the component reset the pristine state incorrectly(?).

References #51