[HeaderBar] Deep link access control
Only show deep links for which the parent app is available.
This will show all deep links for the maintenance app
when the see-maintenance-module is available as a user authority.
When the authority is not present the deep links for those apps are
hidden.

For the above to work properly a 'parentApp' property should be
present on the definition of the deep link menu items.

References #66