[DataTable] Style updates
Reduces the height of the table rows to 50px.
Makes the text inside the table cells shorten by ellipsis when the
text becomes too long. The display value is also added as a title
prop to the string values to allow the user to see the full value.
This is not ideal on very small screens < 1100 as the columns become
very small.