Rework value rendering in DataTable (#106)
Specifically related to the date bug (DHIS2-955)
Dates will now be displayed using the Intl.dateFormat (https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/DateTimeFormat)
When this is not available we'll parse the string and display it.

Additionally the rework makes adding custom renderers for values easier.
The old solution required modification of the Row component to be able
to displayer other values. With the new solution addValueRenderer can be used.