Do no rely on onChange, but use a prop instead
The wrapper cannot really now the implementation of the wrapped
component.
Best is for the wrapped component to pass a function in props which is
called when a user is selected from the list and the input text is
changed.
The function defined in the wrapped component should update the state to
keep the connected input component in sync.