FileMenu - FavoritesDialog integration (#305)
Fixes DHIS2-4482.

Changes proposed in this pull request:

sync data between FileMenu and FavoritesDialog, when rename/delete are performed via FileMenu (ie. refresh the list in FavoritesDialog when is reopened after a rename/delete operation)
remove the actions menu in FavoritesDialog as it makes it difficult to keep a synced state between app, FileMenu and FavoritesDialog.