fix/re-order build script so --watch applies to babel transpiling (#352)
Problem: js changes in period-selector-dialog and core were not re-transpiled when running yarn watch. This is because the --watch option gets applied at the end of the command. When you have a compound npm script (command1 && command2), then the option is applied to the last command. Downside of the change in this commit is that css changes dont trigger the build.