Fix arrow handling in user list for mentions
The sorting of the user list cannot be done only when rendering the
component, the user list in the component's state must have the same
order, otherwise the arrow selection - which is using the state's user
list as index reference - is jumping around the list.
Opted for requesting an ordered list from the API instead, since the
render uses the same object stored in the state.