fix(richtext): skip parsing text in links
If links are detected in the text, skip parsing emphasis and emoji in
them.
A list of link is generated and during parsing the current position of
the marker is checked against the start and end position of links.
If the marker's position falls in any of the links' range, skip parsing.

Fixes DHIS2-6821.