fix: fix interpretation comment delete
Before it was always deleting the last added comment, regardless of
which delete button was clicked.
Fixed by moving the DeleteDialog in the comments list component, we only
need one dialog anyway.
The comment to delete is set in the state via callback when the Delete button
on a comment is used.
When the dialog is confirmed the comment to delete is retrieved from the
state.