fix: upgrade dhis2/analytics (#567)
Upgrade to the latest @dhis/analytics library. No new features or bug fixes. This is an attempt to resolve an issue in dashboards where a seemingly wrong version of ui-core is causing a showstopping error. It seems really unlikely that the version of ui-core in period-selector (via @dhis2/analytics) would be the cause, but keeping dependencies current is usually a good idea anyway, right? :)