fix: custom i18n extract script to insert keys as the default value in pot file [TECH-481] (#653)
cli-app-scripts extract script sets the msgstr = '' in the en.pot file. But for plurals, it is necessary to provide the default (English) plural strings. But Transifex rejects a pot file that is inconsistent (some but not all msgstr filled out). With this change, all msgstr in the pot file will be filled out with value of the key, unless a defaultValue is defined, as it is for plurals (like this: https://github.com/dhis2/d2-ui/blob/master/packages/interpretations/src/components/Interpretation/Replies.js#L77)

The approach was to take the existing extract script from cli-app-scripts, and provide a function for the defaultValue option (lines 44-45 of ./scripts/extract.js).