fix: replace unsafe componentWillReceiveProps method
Use componentDidMount instead, so that the data is fetched also when the
component is used in a conditional rendering situation, like in the new
File Menu component which uses @dhis2/ui components.
In that case, both open and refreshData props need to be passed with a
true value.
The old componentWillReceiveProps tho, never triggered, because those
props don't change in that case, they are just set when the Favorites
component renders.
componentDidMount is executed after the component is rendered, so the
condition for open and refreshData is checked also when using the
component without toggling the open prop.
componentDidUpdate is executed when a prop changes, this is triggered
when using the old approach of toggling the open prop.